migrate some file system methods to FileSystemDeposit
